add a new function: "finding_leap_year_gregorian_calendar"
A leap year occurs under the following conditions:

The year is evenly divisible by 4.

If the year is evenly divisible by 100, it must also be evenly divisible by 400.

In other words, a leap year is:

Any year that is evenly divisible by 4 but not divisible by 100, or

Any year that is evenly divisible by both 100 and 400.