add a new function: "average_grade_calculator"
This function will get user's grades as long as they have not entered (-1), which means to exit the function and print the final average result. There is also a limit to the number of grades the user can enter, in this case it has been set to (50) but you can change or remove it.