add a new function: "sumOfDivisibleNumbers_2"
This function calculates and prints the sum of numbers that are divisible by either 3 or 5 up to a user-defined limit. It begins by prompting the user to enter an integer number. The function offers two approaches: an inefficient solution that checks each number in the range individually, and an optimized solution that iterates through multiples of 3 and 5 separately. In the optimized solution, the code first iterates through multiples of 3, printing the numbers and adding them to the sum. It then iterates through multiples of 5, ensuring not to double-count numbers already considered (i.e., multiples of both 3 and 5). Finally, the sum of all divisible numbers is printed.