add a new function: "sumOfOddNumbersInASequence"
This function calculates the sum of all odd numbers within a user-provided sequence. It begins by prompting the user to enter the number of values (n) in the sequence. The function then iterates through the sequence, asking the user to input each value. For each value, the function checks if it is odd (i.e., not divisible by 2). If the value is odd, it is added to the oddSum. After processing all the values, the function prints the sum of the odd numbers.