add two new functions: "sumOfDigits_recursive" and "sumOfDigits_recursiveApproach"
This program computes the sum of the digits of a number recursively.

Recursive Function: sumOfDigits_recursiveApproach

This function calculates the sum of the digits of the input number num.

Base Case:

If num is a single-digit number (i.e., num <= 9), it returns the number itself.

Recursive Case:

It computes the last digit of num using num % 10 and recursively adds it to the result of the function called on the rest of the digits (num / 10).

Helper Function: sumOfDigits_recursive

It defines the input number num (hardcoded as 953519) and uses the recursive function to compute the sum of its digits.